Fixes and rename package
