+++ rapids-4-spark-tools/src/main/scala/com/nvidia/spark/rapids/tool/profiling/ProfileUtils.scala
@@ -78,3 +78,4 @@ object ProfileUtils {
     ArrayBuffer(FileUtil.stat2Paths(allStatus): _*)
   }
 }
+
