+++ rapids-4-spark-tools/src/main/scala/com/nvidia/spark/rapids/tool/qualification/QualificationArgs.scala
@@ -13,38 +13,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.nvidia.spark.rapids.tool.profiling
+package com.nvidia.spark.rapids.tool.qualification
 
 import org.rogach.scallop.{ScallopConf, ScallopOption}
 
 class QualificationArgs(arguments: Seq[String]) extends ScallopConf(arguments) {
 
   banner("""
-Spark qualification tool
+RAPIDS Accelerator for Apache Spark qualification tool
 
 Example:
 
 # Input 1 or more event logs from local path:
-./bin/spark-submit --class com.nvidia.spark.rapids.tool.profiling.QualificationMain
-<Spark-Rapids-Repo>/rapids-4-spark-tools/target/rapids-4-spark-tools-<version>.jar
- /path/to/eventlog1 /path/to/eventlog2
+./bin/spark-submit --class com.nvidia.spark.rapids.tool.qualification.QualificationMain
+rapids-4-spark-tools_2.12-<version>.jar /path/to/eventlog1 /path/to/eventlog2
 
 # Specify a directory of event logs from local path:
-./bin/spark-submit --class com.nvidia.spark.rapids.tool.profiling.QualificationMain
-<Spark-Rapids-Repo>/rapids-4-spark-tools/target/rapids-4-spark-tools-<version>.jar
- /path/to/DirOfManyEventLogs
+./bin/spark-submit --class com.nvidia.spark.rapids.tool.qualification.QualificationMain
+rapids-4-spark-tools_2.12-<version>.jar /path/to/DirOfManyEventLogs
 
 # If any event log is from S3:
 export AWS_ACCESS_KEY_ID=xxx
 export AWS_SECRET_ACCESS_KEY=xxx
-./bin/spark-submit --class com.nvidia.spark.rapids.tool.profiling.QualificationMain
-<Spark-Rapids-Repo>/rapids-4-spark-tools/target/rapids-4-spark-tools-<version>.jar
-s3a://<BUCKET>/eventlog1 /path/to/eventlog2
+./bin/spark-submit --class com.nvidia.spark.rapids.tool.qualification.QualificationMain
+rapids-4-spark-tools_2.12-<version>.jar s3a://<BUCKET>/eventlog1 /path/to/eventlog2
 
 # Change output directory to /tmp
-./bin/spark-submit --class com.nvidia.spark.rapids.tool.profiling.QualificationMain
- <Spark-Rapids-Repo>/rapids-4-spark-tools/target/rapids-4-spark-tools-<version>.jar
- -o /tmp /path/to/eventlog1
+./bin/spark-submit --class com.nvidia.spark.rapids.tool.qualification.QualificationMain
+rapids-4-spark-tools_2.12-<version>.jar -o /tmp /path/to/eventlog1
 
 For usage see below:
     """)
